修正userid

+++ musicapi/src/main/java/com/cyl/musicapi/netease/LoginInfo.kt
@@ -6,41 +6,41 @@ import com.google.gson.annotations.SerializedName
 data class Account(@SerializedName("salt")
                    val salt: String = "",
                    @SerializedName("vipType")
-                   val vipType: Long = 0,
+                   val vipType: Int = 0,
                    @SerializedName("userName")
                    val userName: String = "",
                    @SerializedName("type")
-                   val type: Long = 0,
+                   val type: Int = 0,
                    @SerializedName("ban")
-                   val ban: Long = 0,
+                   val ban: Int = 0,
                    @SerializedName("anonimousUser")
                    val anonimousUser: Boolean = false,
                    @SerializedName("createTime")
                    val createTime: Long = 0,
                    @SerializedName("tokenVersion")
-                   val tokenVersion: Long = 0,
+                   val tokenVersion: Int = 0,
                    @SerializedName("id")
                    val id: Long = 0,
                    @SerializedName("whitelistAuthority")
-                   val whitelistAuthority: Long = 0,
+                   val whitelistAuthority: Int = 0,
                    @SerializedName("baoyueVersion")
-                   val baoyueVersion: Long = 0,
+                   val baoyueVersion: Int = 0,
                    @SerializedName("viptypeVersion")
-                   val viptypeVersion: Long = 0,
+                   val viptypeVersion: Int = 0,
                    @SerializedName("donateVersion")
-                   val donateVersion: Long = 0,
+                   val donateVersion: Int = 0,
                    @SerializedName("status")
-                   val status: Long = 0)
+                   val status: Int = 0)
 
 
 data class LoginInfo(@SerializedName("clientId")
                      val clientId: String = "",
                      @SerializedName("code")
-                     val code: Long = 0,
+                     val code: Int = 0,
                      @SerializedName("msg")
                      val msg: String = "",
                      @SerializedName("loginType")
-                     val loginType: Long = 0,
+                     val loginType: Int = 0,
                      @SerializedName("profile")
                      val profile: Profile,
                      @SerializedName("bindings")
@@ -52,7 +52,7 @@ data class LoginInfo(@SerializedName("clientId")
 
 
 data class BindingsItem(@SerializedName("expiresIn")
-                        val expiresIn: Long = 0,
+                        val expiresIn: Int = 0,
                         @SerializedName("expired")
                         val expired: Boolean = false,
                         @SerializedName("tokenJsonStr")
@@ -62,7 +62,7 @@ data class BindingsItem(@SerializedName("expiresIn")
                         @SerializedName("id")
                         val id: Long = 0,
                         @SerializedName("type")
-                        val type: Long = 0,
+                        val type: Int = 0,
                         @SerializedName("userId")
                         val userId: Long = 0,
                         @SerializedName("url")
@@ -76,9 +76,9 @@ data class Profile(@SerializedName("detailDescription")
                    @SerializedName("backgroundUrl")
                    val backgroundUrl: String = "",
                    @SerializedName("gender")
-                   val gender: Long = 0,
+                   val gender: Int = 0,
                    @SerializedName("city")
-                   val city: Long = 0,
+                   val city: Int = 0,
                    @SerializedName("signature")
                    val signature: String = "",
                    @SerializedName("description")
@@ -86,7 +86,7 @@ data class Profile(@SerializedName("detailDescription")
                    @SerializedName("remarkName")
                    val remarkName: String? = null,
                    @SerializedName("accountStatus")
-                   val accountStatus: Long = 0,
+                   val accountStatus: Int = 0,
                    @SerializedName("avatarImgId")
                    val avatarImgId: Long = 0,
                    @SerializedName("defaultAvatar")
@@ -96,17 +96,17 @@ data class Profile(@SerializedName("detailDescription")
                    @SerializedName("backgroundImgIdStr")
                    val backgroundImgIdStr: String = "",
                    @SerializedName("province")
-                   val province: Long = 0,
+                   val province: Int = 0,
                    @SerializedName("nickname")
                    val nickname: String = "",
                    @SerializedName("djStatus")
-                   val djStatus: Long = 0,
+                   val djStatus: Int = 0,
                    @SerializedName("avatarUrl")
                    val avatarUrl: String = "",
                    @SerializedName("authStatus")
-                   val authStatus: Long = 0,
+                   val authStatus: Int = 0,
                    @SerializedName("vipType")
-                   val vipType: Long = 0,
+                   val vipType: Int = 0,
                    @SerializedName("followed")
                    val followed: Boolean = false,
                    @SerializedName("userId")
@@ -114,10 +114,10 @@ data class Profile(@SerializedName("detailDescription")
                    @SerializedName("mutual")
                    val mutual: Boolean = false,
                    @SerializedName("authority")
-                   val authority: Long = 0,
+                   val authority: Int = 0,
                    @SerializedName("backgroundImgId")
                    val backgroundImgId: Long = 0,
                    @SerializedName("userType")
-                   val userType: Long = 0)
+                   val userType: Int = 0)
